requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
pydantic-settings
